+++ models/devicetypeobject.go
@@ -1,7 +1,7 @@
 package models
 
 type DeviceTypeObject struct {
-	DeviceTypeId *int    `json:"DeviceTypeId,omitempty"`
+	DeviceTypeId *int64  `json:"DeviceTypeId,omitempty"`
 	Name         *string `json:"Name,omitempty"`
 	Referral     *string `json:"Referral,omitempty"`
 }
