+++ models/frontriskfreebets.go
@@ -1,7 +1,7 @@
 package models
 
 type FrontRiskFreeBets struct {
-	BonusId     *int                 `json:"BonusId,omitempty"`
+	BonusId     *int64               `json:"BonusId,omitempty"`
 	Status      *RiskFreeBetStatuses `json:"Status,omitempty"`
 	DateClaimed *IGCTime             `json:"DateClaimed,omitempty"`
 }
