+++ models/gamecategoryfront.go
@@ -1,8 +1,8 @@
 package models
 
 type GameCategoryFront struct {
-	GameId       *int    `json:"GameId,omitempty"`
-	CategoryId   *int    `json:"CategoryId,omitempty"`
+	GameId       *int64  `json:"GameId,omitempty"`
+	CategoryId   *int64  `json:"CategoryId,omitempty"`
 	CategoryName *string `json:"CategoryName,omitempty"`
-	OrderNumber  *int    `json:"OrderNumber,omitempty"`
+	OrderNumber  *int64  `json:"OrderNumber,omitempty"`
 }
