+++ models/gamejackpot.go
@@ -1,7 +1,7 @@
 package models
 
 type GameJackpot struct {
-	GameId         *int     `json:"GameId,omitempty"`
+	GameId         *int64   `json:"GameId,omitempty"`
 	GameName       *string  `json:"GameName,omitempty"`
 	Vendor         *string  `json:"Vendor,omitempty"`
 	CurrencyFormat *string  `json:"CurrencyFormat,omitempty"`
