+++ models/partialuserdata.go
@@ -1,8 +1,8 @@
 package models
 
 type PartialUserData struct {
-	UserID               *int    `json:"UserID,omitempty"`
+	UserID               *int64  `json:"UserID,omitempty"`
 	IBAN                 *string `json:"IBAN,omitempty"`
-	BankID               *int    `json:"BankID,omitempty"`
+	BankID               *int64  `json:"BankID,omitempty"`
 	SocialSecurityNumber *string `json:"SocialSecurityNumber,omitempty"`
 }
