+++ models/payrequestmodel.go
@@ -1,8 +1,8 @@
 package models
 
 type PayRequestModel struct {
-	PaymentMethodId *int     `json:"PaymentMethodId"`           // Always required. The PaymentMethodId should be retrieved by calling GetUserDepositMethods or GetUserWithdrawMethods
-	PaymentTypeId   *int     `json:"PaymentTypeId"`             // Always required. The PaymentTypeId should be retrieved by calling GetUserDepositMethods or GetUserWithdrawMethods
+	PaymentMethodId *int64   `json:"PaymentMethodId"`           // Always required. The PaymentMethodId should be retrieved by calling GetUserDepositMethods or GetUserWithdrawMethods
+	PaymentTypeId   *int64   `json:"PaymentTypeId"`             // Always required. The PaymentTypeId should be retrieved by calling GetUserDepositMethods or GetUserWithdrawMethods
 	Amount          *float64 `json:"Amount"`                    // Always required. The amount the users has requested
 	BonusCode       *string  `json:"BonusCode,omitempty"`       // Optional. The Bonus Code, if any
 	LanguageAlpha2  *string  `json:"LanguageAlpha2,omitempty"`  // Optional. The Language Alpha2 the user has choosen. If none passed, the default user language will be used
@@ -13,7 +13,7 @@ type PayRequestModel struct {
 	CreditCard      *string  `json:"CreditCard,omitempty"`      // Required only for IGC Credit Card Deposit. This is the credit card number. Not required for EPG
 	ExpiryDate      *string  `json:"ExpiryDate,omitempty"`      // Required only for IGC Credit Card Deposit. This is the credit card expiry date. Not required for EPG
 	CV2             *string  `json:"CV2,omitempty"`             // Required only for IGC Credit Card Deposit. This is the credit card cv2. Not required for EPG
-	UserCardId      *int     `json:"UserCardId,omitempty"`      // Required only for IGC or EPG Credit Card Withdrawals. This is mandatory for withdrawals and should be available from the meta data api call.
+	UserCardId      *int64   `json:"UserCardId,omitempty"`      // Required only for IGC or EPG Credit Card Withdrawals. This is mandatory for withdrawals and should be available from the meta data api call.
 	PayeeName       *string  `json:"PayeeName,omitempty"`       // Required only for Bank Withdrawl. This is the Customer name to be shown in the bank withdrawal transaction
 	CountryCode     *string  `json:"CountryCode,omitempty"`     // Required only for Bank Withdrawl. This is the Country code to be shown in the bank withdrawal transaction
 	AccountNumber   *string  `json:"AccountNumber,omitempty"`   // Required only for Bank Withdrawl. This is the Account Number to be shown in the bank withdrawal transaction
