+++ models/publicbonustypeobject.go
@@ -1,7 +1,7 @@
 package models
 
 type PublicBonusTypeObject struct {
-	TypeId    *int    `json:"TypeId,omitempty"`
-	ProductId *int    `json:"ProductId,omitempty"`
+	TypeId    *int64  `json:"TypeId,omitempty"`
+	ProductId *int64  `json:"ProductId,omitempty"`
 	Type      *string `json:"Type,omitempty"`
 }
