+++ models/securityquestion.go
@@ -1,6 +1,6 @@
 package models
 
 type SecurityQuestion struct {
-	ID       *int    `json:"ID,omitempty"`
+	ID       *int64  `json:"ID,omitempty"`
 	Question *string `json:"Question,omitempty"`
 }
