+++ models/sendemailmodel.go
@@ -1,7 +1,7 @@
 package models
 
 type SendEmailModel struct {
-	EmailTemplateID            *int               `json:"EmailTemplateID,omitempty"`
+	EmailTemplateID            *int64             `json:"EmailTemplateID,omitempty"`
 	SendEmailTo                *string            `json:"SendEmailTo,omitempty"`
 	EmailBodyPlaceholderValues *map[string]string `json:"EmailBodyPlaceholderValues,omitempty"`
 	LanguageCode               *string            `json:"LanguageCode,omitempty"`
